refactor: split storybook tools helpers into individual files
