+++ packages/storybook-tools-helpers/src/index.ts
@@ -1,1 +1,3 @@
-export * from './lib/storybook-tools-helpers';
+export * from './lib/getListTemplate';
+export * from './lib/getOptionsArgTypes';
+export * from './lib/getTemplate';
