+++ packages/storybook-tools-helpers/src/lib/getOptionsArgTypes.tsx
@@ -1,0 +1,7 @@
+export const getOptionsArgTypes = (options: string[]) => ({
+  description: '**options:**',
+  table: {
+    type: { summary: options.map((option) => `'${option}'`).join('|') },
+  },
+  control: { type: 'select', options },
+});
